ExcimerClient: Explicitly cast CURLOPT_TIMEOUT_MS value to integer
It appears to work fine without, but the option is documented as taking
an integer ("Long" type in C), so let's pass that explicitly instead.

Per Aaron's CR at
<https://gerrit.wikimedia.org/r/c/operations/mediawiki-config/+/902529/4/>.

+++ src/ExcimerClient.php
@@ -52,6 +52,8 @@ class ExcimerClient {
 	 *       parameter is passed
 	 *   - period: The sampling period in seconds (default 0.001)
 	 *   - timeout: The request timeout for ingestion requests, in seconds.
+	 *     Use zero to enforce no timeout (default 0).
+	 *     Values smaller than 0.001 (1ms) are not supported.
 	 *   - hashKey: A secret key to be included in the hash, when mapping
 	 *     request IDs to profile IDs. Defaults to no key.
 	 *   - profileIdLength: The number of hexadecimal characters in a generated
@@ -317,7 +319,7 @@ class ExcimerClient {
 		curl_setopt_array( $ch, [
 			CURLOPT_POSTFIELDS => $data,
 			CURLOPT_USERAGENT => 'ExcimerUI',
-			CURLOPT_TIMEOUT_MS => $this->config['timeout'] * 1000,
+			CURLOPT_TIMEOUT_MS => (int)( $this->config['timeout'] * 1000 ),
 			CURLOPT_RETURNTRANSFER => true,
 		] );
 		$result = curl_exec( $ch );
